0x7

+++ 0x17-doubly_linked_lists/lists.h
@@ -1,5 +1,6 @@
 #ifndef LISTS_H
 #define LISTS_H
+#include <stdlib.h>
 /**
  * struct dlistint_s - doubly linked list
  * @n: integer
@@ -14,6 +15,6 @@ typedef struct dlistint_s
 	struct dlistint_s *prev;
 	struct dlistint_s *next;
 } dlistint_t;
-int print_dlistint(const dlistint_t *h);
+size_t print_dlistint(const dlistint_t *h);
 #endif
